fixed issue with removal of duplicate permissions in mysql environments

+++ update/build-1430.pl
@@ -52,10 +52,20 @@ sub _RemoveDuplicatePermissions {
         last;
     }
 
+    # prepare temp table
+    if ( !$DBObject->Do( SQL => "CREATE TABLE tmp_role_permission AS SELECT * FROM role_permission" ) ) {
+        $LogObject->Log(
+            Priority => 'error',
+            Message  => "Unable to create temporary permission table for deletion!"
+        );
+        return;
+    }
+
+    # do the actual deletion
     my $Result = $DBObject->Do(
         SQL => 'DELETE FROM role_permission rp 
                  WHERE EXISTS (
-                     SELECT id FROM role_permission rp2 
+                     SELECT id FROM tmp_role_permission rp2 
                       WHERE rp2.role_id = rp.role_id 
                         AND rp2.target = rp.target 
                         AND rp2.type_id = rp.type_id
